// src/redux/Trips/models/TripEntity.ts
export interface TripEntity {
  endStationId: number;
  endStationLatitude: number;
  endStationLongitude: number;
  endStationName: string;
  gender: number;
  nameLocalizedValue0: string;
  startStationId: number;
  startStationLatitude: number;
  startStationLongitude: number;
  startStationName: string;
  tripduration: number;
  starttime: string;
  stoptime: string;
}